// src/drawing/types.ts
export type DrawBoxOptions = {
  lineWidth?: number
  color?: string
}

export type DrawTextOptions = {
  lineWidth?: number
  fontSize?: number
  fontStyle?: string
  color?: string
}

export type DrawLandmarksOptions = {
  lineWidth?: number
  color?: string
}

export type DrawOptions = {
  lineWidth: number
  fontSize: number
  fontStyle: string
  color: string
}